Modernizes with Stow
TL;DR
-----

Switch to GNU Stow to manage links and tidy up

Details
-------

Updates this respository to use GNU Stow instead of a script that
sets up a bunnch of symobolic links. This makes the repo a bit
tidier to work with and avoids having a bunch of symlink related
logic in the bootstrap script called by Strap.

I also took the opportunity to set up my usual project structure
with encrypted secrets and some other ergonomic features. This
will help me get rid of secrets in the repository and make it
public in the future.

Lastly I cleaned up quite a few things that I'm not using.

To Do
-----

* Add secrets into the `secrets` directory
* Captures more dotfiles and application preferences