Converts color palette links to hard links
TL;DR
-----

Uses hard links instead of symlinks for color palette files

Details
-------

Apples color picker wasn't following the symlinks in that Stow
created in `~/Library/Colors`, but would work if I copied the files
over from the repository. I decided to try out hard links as that
seemed like the solution to the problem.

Since the hard links worked, this change updates the after setup
script to convert to symlinks Stow creates for the `palettes`
module into hard links. The update is in that script since it
depends on the behavior of the GNU `readlink` which is installed
as part of `coreutilis` as `greadlink`.

I also renamed the Solarized panel to have a capital S for
consistency with other palettes, and added a new one I'm going
to be needing soon.