Supports encrypted secrets to keep things safe (#7)
TL;DR
-----

Adds scaffolding to take care of secrets and usese it for AWS
credentials

Details
-------

Uses an improved version of my `secrets` handling script from
[my model repository](https://github.com/crdant/model) to allow for
encrypted secrets in this repository. This will let me make it a
public repo which will both simplify using it and allow others
to share in the wealth.

The `secrets` script is updated to ensure sane permissions on the
decrypted secret (0x600) and to properly handle subdirectories
under `${SECRETS_DIR}`. The latter change allowed me to make a
second call to `stow` using the secrets directory to lay in both
sensitive and non-sensitive files. I had to make a slight change
to how I'm using Stow to account for this layering. Instead of
allowing Stow to automatically fold/unfold, I had to explicitly
direct it not to fold with `--no-folding`.

While testing these changes, I noticed that color palette files
weren't the only types of files that needed to be hard links. I
also realized that the hard links for the color palettes blocked
idempotency. I made corrections so that property list files in
`~/Library/Preferences` are also converted to hard links, and to
back out the hardlinks before attempting to setup the symlink
farm again.