Removes troublesome preference files (#19)
TL;DR
-----

Deltes preferences files that caused issues with bootstrapping

Details
-------

Gets rid of two packaes that were causing too many issues in terms
of getting a bootstrap to run cleanly: `global` and `terminal`. The
likely causse was the files being overwritten during the time it
took to link them and to convert the links to hard links.