Puts Darwin-specifc ZSH config in the right place (#21)
TL;DR
-----

Moves `zshrc_darwin` file to the place it's expected by the import
in `.zshrc`

Details
-------

Corrects an error where the file `zshrc_darwin` was not found when
`.zshrc` includes it. The base file needed to be in the right
subdirectory so that when the link farm is set up it was linked at
the right spot.