Corects for missing `mash` jump box SSH config (#50)
TL;DR
-----

Updates SSH config to use `mash` host as a jumpbox

Details
-------

In the big change that added configuration for a jumpbox I omitted some
SSH configuration chnages for using the `mash` host. This change
includes those configurations that enable the same SSH configurations
that I need for a jumpbox for the host `mash`. It also uses the `mash`
hostname for the TMUX-based SSH connections named `window` and
`fullscreen`.