Migrates brew installss to Nix packages (#59)
TL;DR
-----

Uses Nix packages for things that were installed as brews

Details
-------

I did a quick review of the brews I was installing via Homebrew and
realized that for all for each of them either a package was available or
I dind't really use it. This change migrates to the pacakge for the ones
that I could and deletes the others.