Adds a brew for `llm` (#61)
TL;DR
-----

Usess Homebrew to install the `llm` CLI

Details
-------

I was trying to install the `llm` CLI using Nix, which (properly) does
not allow installing plugins with the tool. I got a good way toward
setting up plugins and installing them using the mechanism it provides,
but it's pretty difficult with the way Nix install Python modules.

This change acknowledges that I needed to move on and start working with
the tool instead of playing with getting a strong, repetable install. It
installs `llm` with Homebrew with allows/requires installing the plugins
with the command itself.