Moves the Nix Store to its own disk
TL;DR
-----

Splits Nix Store and root to their own disks

Details
-------

Disk space was becoming a problem with the regular Nix updates, an that
was leading to issues with a full root disk. This change reconfigures
Nix to use a second disk for the store to protect the root disk and
manage the space independently.