Re-enables documentation on Darwin (#80)
TL;DR
-----

Turns documentation back on for Nix Darwin configuration

Details
-------

There was a bug for a while where having documentation enabled in the
Nix Darwin configuration would cause the build to fail. I had turned if off
to move forward. I tried turning it on for my most recent builds to see
that happened.

This changes persists the return to enabling documentation.