Installs Ghosty as a cask
TL;DR
-----

Removes attempt to install a Ghostty brew and uses the cask

Detiails
--------

Corrects Darwin configuration error where the `ghostty` cask was
referenced as a brew instead of a cask.  While I was add it, I switched
to use the "tip" version for the very latest version of Ghosty.