Uses a new version of Claude Code
TL;DR
-----

Updatses the version of Claude Code in my home environment

Details
-------

Adds an overlay to my home environment that bumps the version of
Claude Code to a newer version that the one currently available in Nix
unstable.

Alongside the version bump, updates my home environment to workaround
a bug in Cluade Code where it will not load slash commands using
symlinks. Currently there's a hacky workaround in place that copies
the slash comand content to the right place rather than linking to
them in the Nix store. That hack is sad, but it needs to be there
until Anthropic fixes their code.