Adds my MBTA MCP server to my home environment (#117)
TL;DR
-----

Install `mbta-mcp-server` for my user accounts and tools

Details
-------

Incorporates the MCP server that Claude and wrote for the accessing
MBTA data into my home environment. Includes a Nix package to install
the MCP server, uses that package in my home environment, and
configures Goose and Claude desktop to use it.