Corrects state verison conditional (#149)
TL;DR
-----

Fixes the logic for determine how to specify the state version

Details
-------

Cleans up the logic for determining the state version. The earlier
logic checked for the `defaults` attribute in the wrong place so it
alwyas used the string that NixOS expects.

Alos fixes the conditional for including MacOS defaults to check the
right place and moves the state version to the base module from the
system defaults module.