Moves the GitHub CLI to the development module
TL;DR
-----

Shifts the GitHub CLI from the CI/CD module to the development module

Details
-------

Locates the GitHub CLI and its configuration int the development
module. It was formerly in the CI/CD module and that doesn't really
make sense since it's used for a lot more than that.