Temporarily(?) Comments out incompatible `llm` tool
TL;DR
-----

Avoids issues with install the `llm` CLI by avoiding installing RAG
tool

Details
-------

Skips installing the RAG plugin for the `llm` CLI. The plugin provides
a tool for RAG to the `llm` CLI, but it has a dependency on an older
version of `llm` than the one we're installing. Hopefully, they
matintainer will bump that up and we can install it again soon.