Relocates activation script config (#192)
TL;DR
-----

Fixes defect where activation script is defined at the wrong place

Details
-------

Corrects the base home environment module to included that activation
script in the home configuration where it belongs. It was previously
at the same level as the `home` config, but it's supposed to be part
of it.