Refreshes a few things (#198)
TL;DR
-----

Lightly updates a few packages

Details
-------

Updates our flake, Tailscale, Claude Code, and KOTS. This required
adding a variant function to build Go packages with the 1.25 line and
surfaced a couple of other things I needed to deal with: the OPA CLI
was struggling to build and I had a few packages that were Replicated
specific tied to the Kubernetee module in my Home Manager
configuration. All are addressed.